requirements.txt
# Core Deep Learning
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0

# Medical Imaging
monai>=1.0.0
nibabel>=3.2.0
SimpleITK>=2.1.0
pydicom>=2.2.0

# Continual Learning
avalanche-lib>=0.3.0

# Data Processing
pandas>=1.3.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
opencv-python>=4.5.0
albumentations>=1.1.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.8.0
wandb>=0.12.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
h5py>=3.6.0
pillow>=9.0.0

# Evaluation
medpy>=0.4.0
surface-distance>=0.1.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Optional: For faster data loading
# lmdb>=1.3.0
# msgpack>=1.0.0
